Add spectral
