+++ package.json
@@ -4,7 +4,7 @@
   "description": "CLI for rating OpenAPI files",
   "type": "module",
   "scripts": {
-    "build": "tsc --build"
+    "build": "tsc --build && tscp"
   },
   "repository": {
     "type": "git",
@@ -23,7 +23,7 @@
     "rate-my-openapi": "rate-my-openapi.js"
   },
   "lint-staged": {
-    "**/*.{ts,json,md,yml,js,css,html}": [
+    "**/*.{ts,json,md,yml,js,css,html,yaml}": [
       "prettier --write"
     ],
     "**/*.{ts,js}": [
@@ -43,10 +43,15 @@
     "eslint-plugin-unicorn": "^44.0.2",
     "lint-staged": "^13.2.0",
     "prettier-plugin-organize-imports": "^3.2.2",
-    "typescript": "5.0.3"
+    "typescript": "5.0.3",
+    "typescript-cp": "^0.1.8"
   },
   "dependencies": {
     "@quobix/vacuum": "^0.1.5",
+    "@stoplight/spectral-core": "^1.18.0",
+    "@stoplight/spectral-functions": "^1.7.2",
+    "@stoplight/spectral-parsers": "^1.0.2",
+    "@stoplight/spectral-ruleset-bundler": "^1.5.2",
     "chalk": "^5.2.0",
     "dotenv": "^16.1.4",
     "openapi-types": "^12.1.3",
