+++ src/generate-spectral-rating/.spectral.yaml
@@ -1,0 +1,1 @@
+extends: [[spectral:oas, recommended]]
